requirements.txt
pycryptodome==3.18.0
PyYAML==6.0.3
flask==2.3.3
requests==2.32.0
passlib==1.7.4
